refactor: extracted header title to its own component
